Create My Codes
With these codes, I imported the matplotlib.pyplot and seaborn Python libraries used for data visualization. It then calculates the correlation matrix of the data stored in the 'train_2' DataFrame, which is between the pairs of variables in the data. I colored the heatmap using the 'coolwarm' colormap and the numerical values of the correlation coefficients were added to each cell of the heatmap. Finally, I set the title of the drawing to 'Correlation Matrix'.

I used the matplotlib.pyplot Python library to generate histograms for two different columns ('column62' and 'column120'). I set column' and y-axis label to 'Number'.